Extract plan execution limiter
